models/baseline/requirements.txt
# AccessAtlas Baseline Model Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# ML utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# Optional: TensorBoard for advanced logging
# tensorboard>=2.13.0
